interface_modelo/requirements.txt
google-generativeai
streamlit
streamlit-extras
st-pages
streamlit-authenticator
pyyaml
pandas
openpyxl
google
google-cloud-discoveryengine
google-api-core
google-auth
google-cloud-aiplatform
google-cloud-storage